FAT-8657 Added karate tests to cover edge-dcb
