+++ edge-dcb/src/main/resources/karate-config.js
@@ -10,6 +10,8 @@ function fn() {
 
   var config = {
     baseUrl: 'http://localhost:9130',
+    edgeUrl: 'http://localhost:9703',
+    centralServerUrl: 'https://folio-dev-volaris-mock-server.ci.folio.org',
     admin: {tenant: 'diku', name: 'diku_admin', password: 'admin'},
     prototypeTenant: 'diku',
 
@@ -17,67 +19,26 @@ function fn() {
     testAdmin: {tenant: testTenant, name: 'test-admin', password: 'admin'},
     testUser: {tenant: testTenant, name: 'test-user', password: 'test'},
 
-    // define global features
-    login: karate.read('classpath:common/login.feature'),
-    loginRegularUser: karate.read('classpath:common/login.feature'),
-    loginAdmin: karate.read('classpath:common/login.feature'),
-    dev: karate.read('classpath:common/dev.feature'),
-//    variables: karate.read('classpath:global/variables.feature'),
-
-    // define global functions
-    uuid: function () {
-      return java.util.UUID.randomUUID() + ''
-    },
-
-    random: function (max) {
-      return Math.floor(Math.random() * max)
-    },
-
-    randomMillis: function() {
-      return java.lang.System.currentTimeMillis() + '';
-    },
-
-    random_string: function() {
-      var text = "";
-      var possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-      for (var i = 0; i < 5; i++)
-        text += possible.charAt(Math.floor(Math.random() * possible.length));
-      return text;
-    },
-
-    isoDate: function() {
-      // var dtf = java.time.format.DateTimeFormatter.ISO_INSTANT;
-      var dtf = java.time.format.DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss'Z'");
-      var date = java.time.LocalDateTime.now();
-      return dtf.format(date);
-    },
-
-    sleep: function(seconds) {
-      java.lang.Thread.sleep(seconds * 1000);
-    }
+      login: karate.read('classpath:common/login.feature'),
+      loginRegularUser: karate.read('classpath:common/login.feature'),
+      loginAdmin: karate.read('classpath:common/login.feature'),
+      dev: karate.read('classpath:common/dev.feature'),
 
   };
 
-  // Create 20 functions for barcode generation
-  var rand1 = function(i) {
-    karate.set("barcode"+i, function() {
-      return Math.floor(Math.random() * 999999999) + '';
-    });
-  }
-  karate.repeat(20, rand1);
-
-  // Create 70 functions for uuid generation
-  var rand2 = function(i) {
+  // Create 100 functions for uuid generation
+  var rand = function(i) {
     karate.set("uuid"+i, function() {
       return java.util.UUID.randomUUID() + '';
     });
   }
-  karate.repeat(70, rand2);
+  karate.repeat(100, rand);
 
   if (env == 'snapshot-2') {
     config.baseUrl = 'https://folio-snapshot-2-okapi.dev.folio.org:443';
     config.edgeUrl = 'https://folio-snapshot-2.dev.folio.org:8000';
-    config.apikey = 'eyJzIjoic3RhZ2luZ0RpcmVjdG9yIiwidCI6ImRpa3UiLCJ1Ijoic3RhZ2luZ0RpcmVjdG9yIn0';
+    config.apikey = 'eyJzIjoiWHlwaEhYT28wWCIsInQiOiJkaWt1IiwidSI6ImRpa3VfYWRtaW4ifQ==';
+
     config.admin = {
       tenant: 'diku',
       name: 'diku_admin',
@@ -86,26 +47,25 @@ function fn() {
   } else if (env == 'snapshot') {
     config.baseUrl = 'https://folio-snapshot-okapi.dev.folio.org:443';
     config.edgeUrl = 'https://folio-snapshot.dev.folio.org:8000';
-    config.apikey = 'eyJzIjoic3RhZ2luZ0RpcmVjdG9yIiwidCI6ImRpa3UiLCJ1Ijoic3RhZ2luZ0RpcmVjdG9yIn0';
+    config.apikey = 'eyJzIjoiWHlwaEhYT28wWCIsInQiOiJkaWt1IiwidSI6ImRpa3VfYWRtaW4ifQ==';
+
     config.admin = {
       tenant: 'diku',
       name: 'diku_admin',
       password: 'admin'
     }
   } else if (env == 'rancher') {
-       config.baseUrl = 'https://folio-dev-volaris-okapi.ci.folio.org';
-       config.edgeUrl = 'https://folio-dev-volaris-edge-dcb.ci.folio.org';
-       config.apikey = 'eyJzIjoiM0xSbVE4TDUySSIsInQiOiJkaWt1IiwidSI6ImRpa3VfYWRtaW4ifQ==';
-       config.admin = {
-         tenant: 'diku',
-         name: 'diku_admin',
-         password: 'admin'
-       }
-   }
-  else if(env == 'folio-testing-karate') {
+    config.baseUrl = 'https://folio-dev-volaris-okapi.ci.folio.org';
+    config.edgeUrl = 'https://folio-dev-volaris-edge-dcb.ci.folio.org';
+    config.apikey = 'eyJzIjoiWHlwaEhYT28wWCIsInQiOiJkaWt1IiwidSI6ImRpa3VfYWRtaW4ifQ==';
+    config.admin = {
+      tenant: 'diku',
+      name: 'diku_admin',
+      password: 'admin'
+    }
+  } else if(env == 'folio-testing-karate') {
     config.baseUrl = '${baseUrl}';
     config.edgeUrl = '${edgeUrl}';
-    config.apikey = 'eyJzIjoic3RhZ2luZ0RpcmVjdG9yIiwidCI6ImRpa3UiLCJ1Ijoic3RhZ2luZ0RpcmVjdG9yIn0';
     config.admin = {
       tenant: '${admin.tenant}',
       name: '${admin.name}',
@@ -115,9 +75,8 @@ function fn() {
     karate.configure('ssl',true);
   } else if (env != null && env.match(/^ec2-\d+/)) {
     // Config for FOLIO CI "folio-integration" public ec2- dns name
-    config.baseUrl = 'https://folio-snapshot-2-okapi.dev.folio.org:443';
-    config.edgeUrl = 'https://folio-snapshot-2.dev.folio.org:8000';
-    config.apikey = 'eyJzIjoic3RhZ2luZ0RpcmVjdG9yIiwidCI6ImRpa3UiLCJ1Ijoic3RhZ2luZ0RpcmVjdG9yIn0';
+    config.baseUrl = 'http://' + env + ':9130';
+    config.edgeUrl = 'http://' + env + ':8000';
     config.admin = {
       tenant: 'supertenant',
       name: 'admin',
